Refactor rename spawnManager.ts to roomManager.ts as it's more explicit
